+++ data/solutes.py
@@ -1,0 +1,38 @@
+# Solutes database (HSP values)
+
+SOLUTES = {
+        "Piceatannol": {"d": 22.7, "p": 7.11, "h": 26.36, "ro": 20.43},
+        "Resveratrol": {"d": 23.1, "p": 6.1, "h": 20.5, "ro": 16.09},
+        "Curcumin": {"d": 18.2, "p": 8.6, "h": 11.5, "ro": 5.5},
+        "Quercetin": {"d": 21.3, "p": 9.4, "h": 14.2, "ro": 4.8},
+        "Genistein": {"d": 19.0, "p": 8.0, "h": 11.8, "ro": 5.1},
+        "Luteolin": {"d": 20.0, "p": 9.2, "h": 12.7, "ro": 5.0},
+        "Pterostilbene": {"d": 19.2, "p": 6.9, "h": 9.8, "ro": 5.3},
+        "Apigenin": {"d": 19.6, "p": 7.2, "h": 12.1, "ro": 5.2},  # Found in chamomile, parsley
+        "Baicalein": {"d": 18.7, "p": 8.0, "h": 10.5, "ro": 5.3},  # Found in skullcap
+        "Catechin": {"d": 21.0, "p": 9.1, "h": 14.5, "ro": 5.0},  # Found in tea, cocoa
+        "Epicatechin": {"d": 20.8, "p": 8.9, "h": 13.9, "ro": 5.0},  # Found in green tea, cocoa
+        "Hesperetin": {"d": 19.2, "p": 7.6, "h": 10.7, "ro": 5.1},  # Found in citrus fruits
+        "Kaempferol": {"d": 20.2, "p": 9.0, "h": 13.2, "ro": 5.0},  # Found in broccoli, kale
+        "Myricetin": {"d": 21.4, "p": 9.7, "h": 15.0, "ro": 4.9},  # Found in berries, tea
+        "Naringenin": {"d": 19.0, "p": 7.5, "h": 11.0, "ro": 5.2},  # Found in citrus fruits
+        "Rutin": {"d": 21.6, "p": 9.8, "h": 16.3, "ro": 4.8},  # Found in buckwheat, citrus fruits
+        "Taxifolin": {"d": 20.7, "p": 8.8, "h": 13.7, "ro": 5.0},  # Found in milk thistle, onions
+        "Triolein": {"d": 16.4, "p": 3.1, "h": 4.9, "ro": 6.2},
+        "Polystyrene": {"d": 18.6, "p": 4.5, "h": 2.9, "ro": 10.6},
+        "Polyethylene": {"d": 18.0, "p": 0.0, "h": 2.0, "ro": 9.8},
+        "PVC": {"d": 18.2, "p": 7.5, "h": 8.3, "ro": 8.1},
+        "PMMA": {"d": 18.6, "p": 10.5, "h": 7.5, "ro": 8.6},
+        "Nylon-6,6": {"d": 18.6, "p": 5.6, "h": 12.8, "ro": 6.7},
+        "PLA": {"d": 18.6, "p": 9.9, "h": 6.0, "ro": 8.4},
+        "ABS": {"d": 17.6, "p": 8.6, "h": 4.7, "ro": 9.2},
+        "PETG": {"d": 18.2, "p": 6.4, "h": 6.6, "ro": 8.8},
+        "TPU": {"d": 17.5, "p": 9.3, "h": 8.5, "ro": 8.0},
+        "Nylon (PA12)": {"d": 17.0, "p": 3.7, "h": 6.4, "ro": 8.2},
+        "PEEK": {"d": 19.5, "p": 5.2, "h": 5.2, "ro": 10.8},
+        "PVA": {"d": 15.6, "p": 18.8, "h": 16.9, "ro": 7.0},
+        "HIPS": {"d": 18.0, "p": 4.3, "h": 2.7, "ro": 10.4},
+        "ASA": {"d": 18.1, "p": 9.5, "h": 5.0, "ro": 9.0},
+        "PC (Polycarbonate)": {"d": 19.1, "p": 7.5, "h": 6.0, "ro": 9.5},
+        "Pullan": {"d": 18.7, "p": 18.1, "h": 23.5, "ro": 47.7},
+    }
